Some fixes after PR review
